feat: refactor common rust code to utils
