+++ keywise/Cargo.toml
@@ -11,9 +11,10 @@ wasm-opt = false
 crate-type = ["cdylib"]
 
 [dependencies]
+utils = { path = "../utils-rs" }
 serde = { version = "1.0.160", features = ["derive"] }
 serde_json = "1.0.96"
-worker = "0.0.16"
+worker = "0.0.17"
 
 [profile.release]
 lto = true
