+++ keywise/wrangler.toml
@@ -2,10 +2,10 @@ name = "url-shortener"
 main = "build/worker/shim.mjs"
 compatibility_date = "2023-04-27"
 usage_model = "bundled"
-account_id = "b3f9fdfd827152316d080a5ddee59915"
+# account_id = "b3f9fdfd827152316d080a5ddee59915"
 
 kv_namespaces = [
-    { binding = "list", id = "e2ec51da00a54b30b11e843ec6f89957", preview_id = "54ead6b7cd5a48e0ba56c7212dde9c42" }
+    { binding = "list", id = "54b02bab8d394f448dd0c064feab1e86", preview_id = "3a9ba2324e9c4021b76107ead42672e0" }
 ]
 
 [build]
